Phase 4.7: Data Migration to %ProgramData%\EZPos\
- Move database from app folder to %ProgramData%\EZPos\EZPos.db
- Move config from app folder to %ProgramData%\EZPos\config.ini
- Move license from app folder to %ProgramData%\EZPos\license.dat
- Update Database.cs: DbFile now uses CommonApplicationData path
- Update ConfigHelper.cs: config path now uses CommonApplicationData
- Add MigrateToNewDataLocation() in App.xaml.cs to handle one-time migration
- Update Inno Setup script to create ProgramData folders and deploy seed DB there
- Ensures data survives updates and follows Windows app data conventions
- Build: 0 errors, 27 pre-existing warnings

+++ src/Business/Services/StockService.cs
@@ -27,7 +27,7 @@ namespace EZPos.Business.Services
             var record = _store.Products.FirstOrDefault(p => p.Id == productId);
             if (record == null) return false;
 
-            int newStock = record.Stock + changeQty;
+            decimal newStock = record.Stock + changeQty;
             if (newStock < 0) newStock = 0;
 
             ProductRepository.UpdateStock(productId, newStock);
