+++ src/Models/Domain/SaleItem.cs
@@ -5,7 +5,8 @@ namespace EZPos.Models.Domain
         public int Id { get; set; }
         public int SaleId { get; set; }
         public int ProductId { get; set; }
-        public int Quantity { get; set; }
+        /// <summary>Supports decimal for weight-based products (e.g. 1.25 kg).</summary>
+        public decimal Quantity { get; set; }
         public decimal Price { get; set; }
     }
 }
